// extensions/music-assistant-controls/src/music-assistant/external-code/interfaces.ts
export enum MediaType {
  ARTIST = "artist",
  ALBUM = "album",
  TRACK = "track",
  PLAYLIST = "playlist",
  RADIO = "radio",
  AUDIOBOOK = "audiobook",
  PODCAST = "podcast",
  PODCAST_EPISODE = "podcast_episode",
  GENRE = "genre",
  FOLDER = "folder",
  UNKNOWN = "unknown",
}
// ...
export enum RepeatMode {
  OFF = "off", // no repeat at all
  ONE = "one", // repeat current/single track
  ALL = "all", // repeat entire queue
}

export enum PlayerState {
  IDLE = "idle",
  PAUSED = "paused",
  PLAYING = "playing",
}

export enum PlayerType {
  PLAYER = "player", // A regular player.
  GROUP = "group", // A (dedicated) group player or playergroup.
  STEREO_PAIR = "stereo_pair",
}

export enum PlayerFeature {
  POWER = "power",
  VOLUME_SET = "volume_set",
  VOLUME_MUTE = "volume_mute",
  PAUSE = "pause",
  SET_MEMBERS = "set_members",
  MULTI_DEVICE_DSP = "multi_device_dsp",
  SEEK = "seek",
  NEXT_PREVIOUS = "next_previous",
  PLAY_ANNOUNCEMENT = "play_announcement",
  ENQUEUE = "enqueue",
  SELECT_SOURCE = "select_source",
}
// ...
export enum HidePlayerOption {
  NEVER = "never",
  WHEN_OFF = "when_off",
  WHEN_GROUP_ACTIVE = "when_group_active",
  WHEN_SYNCED = "when_synced",
  WHEN_UNAVAILABLE = "when_unavailable",
  ALWAYS = "always",
}
// ...
export interface MediaItemImage {
  type: ImageType;
  path: string;
  provider: string;
  remotely_accessible: boolean;
}
// ...
export interface Artist extends MediaItem {}

export interface Album extends MediaItem {
  year?: number;
  artists: Array<ItemMapping | Artist>;
  album_type: AlbumType;
}

export interface Track extends MediaItem {
  duration: number;
  artists: Array<ItemMapping | Artist>;
  // album track only
  album: ItemMapping | Album;
  disc_number?: number;
  track_number?: number;
}

export interface Playlist extends MediaItem {
  owner: string;
  is_editable: boolean;
}

export interface Radio extends MediaItem {}

export interface Genre extends MediaItem {}

export interface Audiobook extends MediaItem {
  publisher: string;
  authors: string[];
  narrators: string[];
  duration: number;
  fully_played?: boolean;
  resume_position_ms?: number;
}

export interface Podcast extends MediaItem {
  publisher?: string;
  total_episodes?: number;
}

export interface PodcastEpisode extends MediaItem {
  position: number;
  podcast: Podcast | ItemMapping;
  duration: number;
  fully_played?: boolean;
  resume_position_ms?: number;
}

export interface BrowseFolder extends MediaItem {
  path?: string;
  image?: MediaItemImage;
}
export interface RecommendationFolder extends BrowseFolder {
  icon?: string;
  items: MediaItemTypeOrItemMapping[];
}

export type MediaItemType =
  | Artist
  | Album
  | Track
  | Radio
  | Playlist
  | Genre
  | Audiobook
  | Podcast
  | PodcastEpisode
  | BrowseFolder;

export type PlayableMediaItemType = Track | Radio | Audiobook | PodcastEpisode;
// ...
export interface StreamDetails {
  provider: string;
  item_id: string;
  audio_format: AudioFormat;
  media_type: MediaType;
  stream_title?: string;
  duration?: number;

  queue_id?: string;
  fade_in?: boolean;
  loudness?: number;
  loudness_album?: number;
  prefer_album_loudness?: boolean;
  target_loudness?: number;
  volume_normalization_mode?: VolumeNormalizationMode;
  volume_normalization_gain_correct?: number;
  // This contains the DSPDetails of all players in the group.
  // In case of single player playback, dict will contain only one entry.
  dsp?: Record<string, DSPDetails>;
}

// queue_item

export interface QueueItem {
  queue_id: string;
  queue_item_id: string;
  name: string;
  duration: number;
  sort_index: number;
  streamdetails?: StreamDetails;
  media_item?: PlayableMediaItemType;
  image?: MediaItemImage;
  available: boolean;
}

// player_queue

export interface PlayerQueue {
  queue_id: string;
  active: boolean;
  display_name: string;
  available: boolean;
  items: number;
  shuffle_enabled: boolean;
  dont_stop_the_music_enabled: boolean;
  repeat_mode: RepeatMode;
  current_index?: number;
  index_in_buffer?: number;
  elapsed_time: number;
  elapsed_time_last_updated: number;
  state: PlayerState;
  current_item?: QueueItem;
  next_item?: QueueItem;
  radio_source: MediaItemType[];
}

// player

export interface DeviceInfo {
  model: string;
  manufacturer: string;
  software_version?: string;
  model_id?: string;
  manufacturer_id?: string;
  ip_address?: string;
  mac_address?: string;
}

export interface PlayerMedia {
  uri: string; // uri or other identifier of the loaded media
  media_type: MediaType;
  title?: string; // optional
  artist?: string; // optional
  album?: string; // optional
  image_url?: string; // optional
  duration?: number; // optional
  queue_id?: string; // only present for requests from queue controller
  queue_item_id?: string; // only present for requests from queue controller
}

export interface PlayerSource {
  id: string;
  name: string;
  passive: boolean;
  can_play_pause: boolean;
  can_seek: boolean;
  can_next_previous: boolean;
}

export interface PlayerControl {
  control_id: string;
  name?: string;
  [key: string]: any;
}

export interface Player {
  player_id: string;
  provider: string;
  type: PlayerType;
  name: string;
  available: boolean;
  device_info: DeviceInfo;
  supported_features: PlayerFeature[];
  can_group_with: string[];
  enabled: boolean;

  elapsed_time?: number;
  elapsed_time_last_updated?: number;
  current_media?: PlayerMedia;
  state?: PlayerState;
  powered?: boolean;
  volume_level?: number;
  volume_muted?: boolean;
  group_childs: string[];
  active_source?: string;
  source_list: PlayerSource[];
  active_group?: string;
  synced_to?: string;

  group_volume: number;
  display_name: string;
  hide_player_in_ui: HidePlayerOption[];
  icon: string;
  power_control: string;
  volume_control: string;
  mute_control: string;
}